+++ CHANGELOG.md
@@ -24,6 +24,8 @@ All notable changes to this project will be documented in this file.
 
 ### Dependency Upgrades
 
+- Upgrade to Spring Boot 2.7.17. (#120)
+- Upgrade to Spring Dependency Management Plugin 1.1.4. (#120)
 - Upgrade to `jenkins-library` 2.7.4. (#116)
 
 ## [[8.2.0]](https://github.com/iExecBlockchainComputing/iexec-blockchain-adapter-api/releases/tag/v8.2.0) 2023-09-28